chore(Advisor) Clean-up/Refactor
