+++ src/PresentationalComponents/Cards/TotalRiskCard.js
@@ -4,7 +4,6 @@ import './Pathways.scss';
 import {
   Card,
   CardBody,
-  CardFooter,
   CardTitle,
 } from '@patternfly/react-core/dist/esm/components/Card/index';
 import {
@@ -19,16 +18,7 @@ import {
   GridItem,
 } from '@patternfly/react-core/dist/esm/layouts/Grid/index';
 
-import { Text } from '@patternfly/react-core';
-import ArrowRightIcon from '@patternfly/react-icons/dist/esm/icons/arrow-right-icon';
-import CategoryLabel from '../Labels/CategoryLabel';
-import { InsightsLabel } from '@redhat-cloud-services/frontend-components/InsightsLabel';
-import { Link } from 'react-router-dom';
-import { RISK_OF_CHANGE_LABEL } from '../../AppConstants';
 import React from 'react';
-import { RebootRequired } from '../Common/Common';
-import RecommendationLevel from '../Labels/RecommendationLevel';
-import RuleLabels from '../Labels/RuleLabels';
 import chart_color_black_100 from '@patternfly/react-tokens/dist/esm/chart_color_black_100';
 import chart_color_gold_400 from '@patternfly/react-tokens/dist/esm/chart_color_gold_400';
 import chart_color_orange_300 from '@patternfly/react-tokens/dist/esm/chart_color_orange_300';
@@ -37,53 +27,7 @@ import messages from '../../Messages';
 import { strong } from '../../Utilities/intlHelper';
 import { useIntl } from 'react-intl';
 
-const PathwayCard = (props) => {
-  const intl = useIntl();
-  const {
-    name,
-    categories,
-    impacted_systems_count,
-    description,
-    has_incident,
-    reboot_required,
-    slug,
-  } = props;
-
-  return (
-    <Card
-      isFlat
-      isPlain
-      className="adv-c-card-pathway adv__background--global-100"
-    >
-      <CardBody className="body">
-        <Text className="pf-u-pb-sm pf-u-font-weight-bold">{name}</Text>
-        <CategoryLabel key={name} labelList={categories} />{' '}
-        <Link
-          to={`/recommendations/pathways/systems/${slug}`}
-          className="pf-u-font-size-sm"
-        >
-          {intl.formatMessage(messages.topicCardSystemsaffected, {
-            systems: impacted_systems_count,
-          })}
-        </Link>
-      </CardBody>
-      <CardBody className="body pf-u-font-size-sm">{description}</CardBody>
-      <CardBody className="body pf-u-font-size-sm">
-        {has_incident && (
-          <RuleLabels rule={{ tags: 'incident' }} isCompact noMargin />
-        )}{' '}
-        {RebootRequired(reboot_required)}
-      </CardBody>
-      <CardFooter className="footer pf-u-font-size-sm">
-        <Link to={`/recommendations/pathways/${slug}`}>
-          {`${intl.formatMessage(messages.viewPathway)} `}
-          <ArrowRightIcon />
-        </Link>
-      </CardFooter>
-    </Card>
-  );
-};
-const TotalRisk = (props) => {
+export const TotalRiskCard = (props) => {
   const intl = useIntl();
   const {
     impacted_systems_count,
@@ -185,56 +129,3 @@ const TotalRisk = (props) => {
     </Card>
   );
 };
-const Resolution = (props) => {
-  const intl = useIntl();
-  const { reboot_required, name, resolution_risk } = props;
-
-  return (
-    <Card
-      isFlat
-      isPlain
-      className="adv-c-card-pathway adv__background--global-100 pf-u-h-100 flex-row"
-    >
-      <div className="flex-coloumn">
-        <CardTitle>{intl.formatMessage(messages.resolution)}</CardTitle>
-
-        <div className="flex-row ">
-          <div className="halfWidth">
-            <p className="pf-u-font-weight-bold pf-u-font-size-sm pf-u-pl-lg">
-              {intl.formatMessage(messages.remediation)}
-            </p>
-            <p className="pf-u-font-size-sm pf-u-pl-lg">{name}</p>
-          </div>
-
-          <CardBody className=" pf-u-pl-xl halfWidth">
-            <p className="pf-u-font-weight-bold pf-u-font-size-sm">
-              {intl.formatMessage(messages.riskOfChange)}
-            </p>
-            <InsightsLabel
-              text={RISK_OF_CHANGE_LABEL[resolution_risk.risk]}
-              value={resolution_risk.risk}
-              hideIcon
-              isCompact
-            />
-          </CardBody>
-        </div>
-
-        <CardBody className="body pf-u-font-size-sm">
-          {intl.formatMessage(messages.staticRemediationDesc)}
-        </CardBody>
-        <CardBody className="body">{RebootRequired(reboot_required)}</CardBody>
-      </div>
-
-      <div className="pathwayRight pf-u-p-lg ">
-        <p className="pf-u-font-weight-bold pf-u-font-size-sm">
-          {intl.formatMessage(messages.reclvl)}
-        </p>
-        <div>
-          <RecommendationLevel {...props} />
-        </div>
-      </div>
-    </Card>
-  );
-};
-
-export { PathwayCard, TotalRisk, Resolution };
