+++ src/PresentationalComponents/Common/Tables.js
@@ -78,13 +78,6 @@ export const paramParser = () => {
   );
 };
 
-// create url from options
-export const encodeOptionsToURL = (options) => {
-  return Object.entries(options).reduce((acc, [key, value], index) => {
-    return (acc += index === 0 ? `${key}=${value}` : `&${key}=${value}`);
-  }, '');
-};
-
 // capitalizes text string
 export const capitalize = (string) =>
   string[0].toUpperCase() + string.substring(1);
