+++ src/PresentationalComponents/Labels/CategoryLabel.js
@@ -1,6 +1,6 @@
+/* eslint-disable react/jsx-key */
 import AutomationIcon from '@patternfly/react-icons/dist/esm/icons/automation-icon';
 import CubeIcon from '@patternfly/react-icons/dist/esm/icons/cube-icon';
-import { Label } from '@patternfly/react-core/dist/js/components/Label/index';
 import { LabelGroup } from '@patternfly/react-core/dist/js/components/LabelGroup/LabelGroup';
 import LockIcon from '@patternfly/react-icons/dist/esm/icons/lock-icon';
 import PortIcon from '@patternfly/react-icons/dist/esm/icons/port-icon';
@@ -8,8 +8,9 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import messages from '../../Messages';
 import { useIntl } from 'react-intl';
+import LabelIcon from './LabelIcon';
 
-const CategoryLabel = ({ labelList, isCompact }) => {
+const CategoryLabel = ({ labelList }) => {
   const intl = useIntl();
   const sortedFrequency = (arr) =>
     Object.entries(
@@ -19,43 +20,29 @@ const CategoryLabel = ({ labelList, isCompact }) => {
     labelList.map((label) => label.id)
   );
 
-  const label = (icon, text) => (
-    <Label icon={icon} variant="outline" color="blue" isCompact={isCompact}>
-      {text}
-    </Label>
-  );
-  const labels = (id, isCompact) => {
-    if (id === 1) {
-      return label(
-        <AutomationIcon />,
-        intl.formatMessage(messages.availability)
-      );
-    }
-    if (id === 2) {
-      return label(
-        <LockIcon />,
-        intl.formatMessage(messages.security),
-        isCompact
-      );
-    }
-    if (id === 3) {
-      return label(
-        <CubeIcon />,
-        intl.formatMessage(messages.stability),
-        isCompact
-      );
-    }
-    if (id === 4) {
-      return label(
-        <PortIcon />,
-        intl.formatMessage(messages.performance),
-        isCompact
-      );
-    }
-  };
+  const labels = (id) =>
+    [
+      <LabelIcon
+        icon={<AutomationIcon />}
+        text={intl.formatMessage(messages.availability)}
+      />,
+      <LabelIcon
+        icon={<LockIcon />}
+        text={intl.formatMessage(messages.security)}
+      />,
+      <LabelIcon
+        icon={<CubeIcon />}
+        text={intl.formatMessage(messages.stability)}
+      />,
+      <LabelIcon
+        icon={<PortIcon />}
+        text={intl.formatMessage(messages.performance)}
+      />,
+    ][id];
+
   return (
-    <LabelGroup numLabels={1} isCompact={isCompact}>
-      {sortedFrequencyList.map((id) => labels(Number(id[0]), isCompact))}
+    <LabelGroup numLabels={1} isCompact>
+      {sortedFrequencyList.map((id) => labels(Number(id[0] - 1)))}
     </LabelGroup>
   );
 };
