+++ src/SmartComponents/Recs/DetailsPathways.js
@@ -9,10 +9,8 @@ import {
   PageHeaderTitle,
 } from '@redhat-cloud-services/frontend-components/PageHeader';
 import React, { Suspense, lazy, useEffect, useState } from 'react';
-import {
-  Resolution,
-  TotalRisk,
-} from '../../PresentationalComponents/Cards/Pathways';
+import { TotalRiskCard } from '../../PresentationalComponents/Cards/TotalRiskCard';
+import { ResolutionCard } from '../../PresentationalComponents/Cards/ResolutionCard';
 import {
   Tab,
   TabTitleText,
@@ -157,10 +155,10 @@ const PathwayDetails = () => {
           <Main className="pf-u-pb-0">
             <Grid hasGutter>
               <GridItem sm={12} md={6}>
-                <TotalRisk {...pathway} />
+                <TotalRiskCard {...pathway} />
               </GridItem>
               <GridItem sm={12} md={6}>
-                <Resolution {...pathway} />
+                <ResolutionCard {...pathway} />
               </GridItem>
             </Grid>
           </Main>
