--- src/Utilities/_BarDividedList.scss
@@ -1,3 +1,0 @@
-.adv-verticalDivider {
-  color: var(--pf-global--BorderColor--100)
-}
